+++ src/webviews/react/commit-list/CommitRow.tsx
@@ -92,8 +92,6 @@ function TooltipRefRow({
     );
 }
 
-
-
 function CommitMessageCell({
     message,
     refs,